pyproject.toml
[tool.poetry]
name = "django-test-task"
version = "0.1.0"
description = ""
authors = ["IlyaGrebenschikov <ilushagr22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.13"
django = "4.1"
python-dotenv = "^1.0.1"
mysqlclient = "^2.2.6"
mysql-connector-python = "^9.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
